refactor: split flexsearch.ts into modular components
Split monolithic flexsearch.ts (356 lines) into focused modules:

- flexsearch/cache.ts (33 lines) - localStorage cache operations
- flexsearch/scoring.ts (53 lines) - relevance scoring algorithm
- flexsearch/highlighting.ts (33 lines) - text highlighting & snippets
- flexsearch/index.ts (151 lines) - main search logic & public API

Benefits:
- Better separation of concerns (cache, scoring, highlighting, search)
- Improved testability (each module can be tested independently)
- Easier maintenance (changes to scoring don't affect caching)
- Clearer dependencies (explicit imports between modules)
- All files now under 200 lines

No functional changes - all existing imports continue to work via
index.ts re-exports.

Relates to #179
